+++ sdl3-sys/ChangeLog.md
@@ -1,6 +1,7 @@
 # Changes
 
-- next:
+- 0.6.1:
+    - Update SDL to 3.4.2
     - Add `new` and `value` assoc function/method for groups (ids, enums, etc)
     - impl Display for ids under the `display-impls` feature
     - Add metadata for structs and unions